+++ CHANGES.rst
@@ -7,6 +7,8 @@ astropy-helpers Changelog
 - Removing deprecated _test_compat making astropy a hard dependency for
   packages wishing to use the astropy tests machinery. [#314]
 
+- Update numpydoc to v0.7. [#343]
+
 - Removing unused 'register' command since packages should be uploaded
   with twine and get registered automatically. [#332]